+++ packages/expo-build-properties/src/pluginConfig.ts
@@ -179,6 +179,19 @@ export interface PluginConfigTypeAndroid {
    * @default false
    */
   buildFromSource?: boolean;
+  /**
+   * Override the default `reactNativeArchitectures` list of ABIs to build in **gradle.properties**.
+   *
+   * @see [Android documentation](https://developer.android.com/ndk/guides/abis) for more information.
+   *
+   * @example
+   * ```json
+   * ["arm64-v8a", "x86_64"]
+   * ```
+   *
+   * @default ["armeabi-v7a", "arm64-v8a", "x86", "x86_64"]
+   */
+  buildArchs?: string[];
 }
 
 // @docsMissing
@@ -658,6 +671,7 @@ const schema: JSONSchemaType<PluginConfigType> = {
         },
         enableBundleCompression: { type: 'boolean', nullable: true },
         buildFromSource: { type: 'boolean', nullable: true },
+        buildArchs: { type: 'array', items: { type: 'string' }, nullable: true },
       },
       nullable: true,
     },